+++ frontend/.eslintcache
@@ -1,1 +1,1 @@
-[{"C:\\Users\\karan\\git\\book-project\\frontend\\src\\App.tsx":"1"},{"size":707,"mtime":1608906963571,"results":"2","hashOfConfig":"3"},{"filePath":"4","messages":"5","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"xlb0qo","C:\\Users\\karan\\git\\book-project\\frontend\\src\\App.tsx",[]]
+[{"C:\\Users\\karan\\git\\book-project\\frontend\\src\\login\\App.tsx":"1"},{"size":1661,"mtime":1608916754848,"results":"2","hashOfConfig":"3"},{"filePath":"4","messages":"5","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"xlb0qo","C:\\Users\\karan\\git\\book-project\\frontend\\src\\login\\App.tsx",[]]
