--- frontend/src/App.css
@@ -1,18 +1,0 @@
-.App-header {
-  position: fixed;
-  top: 35%;
-  left: 45%;
-}
-
-#app-name {
-  font-weight: normal;
-  padding-left: 50px;
-}
-
-#standard-basic {
-  width: 290px;
-}
-
-#login {
-  width: 290px;
-}
